Refactor pacakges
