+++ packages/client/package.json
@@ -10,8 +10,8 @@
     "author": "dev@synfutures.com",
     "license": "MIT",
     "dependencies": {
-        "@synfutures/base-plugins": "^1.0.0",
         "@synfutures/logger": "^1.0.0",
+        "@synfutures/utils": "^1.0.0",
         "@synfutures/oyster-sdk": "^0.1.13",
         "amqplib": "^0.10.4"
     },
